New documentation, specs improved
All parts of output_parser.py are commented now.
Some minor issues with the accuracy of the specs that were dis-
covered yesterday were removed.
In detail: the specs are stating that the parser never generates
syntactically invalid output. This was, until recently, not true,
since a data-check in the parse_structure routine of OutputParser
prevented popping the last element in case the tokens of the last
element and the current instruction line don't match, which might
lead to syntactically wrong output, the issue is now removed.
On the other hand, the output for malformatted, thus invalid
files will be inaccurate, duh. Not my fault, the specs are
distributed alongside the project.
Minor improvements in the internal API, not of interest to the
normal user.