Rework of loop recognition
Implemented a more modular approach (context- and not jump-oriented to
recognze loops. The next step, recognition of branches is already
prepared, altough the detailed algorithm is still unclear. This was done
to prevent conflicts between different aspects of the Assembly-Parsing,
since the previous approach worked but was
a) fagile
b) probably conflicting with mechanics introduced in future updates.

Done!