Working on branches
No real chnages to the code, just a new testcase and a few ideas.
The algorithm for branch recognition needs tobe described and
formalized, but it isn't that easy: logical connections etc. will
become a challenge if every special case is parsed on its own. Thus,
a algorithm is needed, that is able to predict which parts of code are
executed exclusively and under which conditions. orking on that.