Switch to graph-based parsing!
Graph implementation in a quite good stage.
This also means that the assembly-parser will soone be modified to use
the new Graph-API, which will ease the implementation of more complex
analysis features. For instance, compiler optimizations shuffling the
code layout (for-/while-loops represented as do-loops with a jump to
the condition etc.) will no longer affect the way the program analyzes
the code, since the analysis is reduced to codeflow analysis.