Finding a scientific base
Found an article on structural analysis of CFG's. The information
gathered there can be used in the given project, solving many issues
and reflecting ideas I already had. So... why not?
This means that we need a spanning tree (DUH! I needed to know that some
time BEFORE!), so the generation of a DFS-Tree is implemented, alongside
a complete binary tree representation. Note: Technically it isn't a
binary tree, since the amount of children a node has isn't limited. It's
just not possible to get a different kind of tree from a CFG ;)