Fixed the OR support!
Happy! Really HAPPY! OR is now (hopefully) supported to full extend.
This issue was due to an implementation detail of the if-then and if-then-else
structures that allowed conitional nodes with multiple entry-edges.
Thus, parts of a if-then-with-or of if-then-else-with-or were recognized as
if-then-else's which rendered the recognition of the aforementioned, more complex
patterns completely imossible. The issue is now fixed.

Apart from that, a new test-case was added to check how well the program deals with
optimizng compilers, but the whole idea falde a bit, since the compiler decided to remove
some loops. My bad, it's an _OptimizinG_ compiler after all.