Bugfixes, tests and lots of frustration.
This is a maintenance release. Some problems have been adressed, as well as an analysis-framework has been created, to
ensure the comfort of the workflow the software needs. Some drafting has been done on how to support more complex conditions,
with some success, but nothing has yet been implemented, algorithms exist in a very general form (on paper). It is possible
that they will be implemented in a future update, maybe I'll first concentrate on other modules that yet have to be written (see org file).
Maybe these algorithms will be first tested on a different branch to ensure nothing gets in it's way, since this is an advanced and complex
fature that will need a lot of effort, including some cuts abd changes in already existing code. Good luck, me!