Added parsing and reduction of complex conditions
Now you can feed any combination of && and || to the compiler,
Iridium should be able to distingush between conditions and
"actual" code. Now we need to work on recognizing in what relationship
the nodes considered part of a condition are, that should be doable.
The advanced-cfg branch is obsolete now. Apart from that, code cleaning
is important and should be done over the winter. This includes an API spec
that makes the method args and return values more interchangeable.
Also, add more security checks!

That's it for now, bye.